requirements.txt
instagrapi==2.0.0
pyTelegramBotAPI==4.14.0
ipython==8.17.2
pydantic==1.10.9
PyYAML==6.0.1
Pillow>=10.1.0